Add preview configuration option to suppress ingestion sampling

+++ agent/agent-tooling/src/main/java/com/microsoft/applicationinsights/agent/internal/sampling/AiSampler.java
@@ -27,24 +27,32 @@ import javax.annotation.Nullable;
 // * adds item count to span attribute if it is sampled
 public class AiSampler implements Sampler {
 
+  private final boolean suppressIngestionSampling;
   private final boolean localParentBased;
   private final SamplingPercentage requestSamplingPercentage;
   // when localParentBased=false, then this applies to all dependencies, not only parentless
   private final SamplingPercentage parentlessDependencySamplingPercentage;
-  private final Cache<Long, SamplingResult> recordAndSampleWithItemCountMap = Cache.bounded(100);
+  private final Cache<Double, SamplingResult> recordAndSampleWithSampleRateMap = Cache.bounded(100);
 
   public AiSampler(
       SamplingPercentage requestSamplingPercentage,
-      SamplingPercentage parentlessDependencySamplingPercentage) {
-    this(requestSamplingPercentage, parentlessDependencySamplingPercentage, true);
+      SamplingPercentage parentlessDependencySamplingPercentage,
+      boolean suppressIngestionSampling) {
+    this(
+        requestSamplingPercentage,
+        parentlessDependencySamplingPercentage,
+        suppressIngestionSampling,
+        true);
   }
 
   public AiSampler(
       SamplingPercentage requestSamplingPercentage,
       SamplingPercentage parentlessDependencySamplingPercentage,
+      boolean suppressIngestionSampling,
       boolean localParentBased) {
     this.requestSamplingPercentage = requestSamplingPercentage;
     this.parentlessDependencySamplingPercentage = parentlessDependencySamplingPercentage;
+    this.suppressIngestionSampling = suppressIngestionSampling;
     this.localParentBased = localParentBased;
   }
 
@@ -76,24 +84,29 @@ public class AiSampler implements Sampler {
               : parentlessDependencySamplingPercentage.get();
     }
 
-    if (sp == SamplingPercentage.USE_INGESTION_SAMPLING) {
-      return SamplingResult.recordAndSample();
-    }
-
     if (sp == 0) {
       return SamplingResult.drop();
     }
 
-    if (!shouldRecordAndSample(traceId, sp)) {
+    if (sp != 100 && !shouldRecordAndSample(traceId, sp)) {
       return SamplingResult.drop();
     }
 
-    // sp cannot be 0 here
-    long itemCount = Math.round(100.0 / sp);
-    SamplingResult samplingResult = recordAndSampleWithItemCountMap.get(itemCount);
+    if (sp == 100 && !suppressIngestionSampling) {
+      return SamplingResult.recordAndSample();
+    }
+
+    if (sp == 100) {
+      // ingestion sampling is applied when sample rate is 100 (or missing)
+      // so we set it to 99.99 which will bypass ingestion sampling (and will still be stored as
+      // item count 1)
+      sp = 99.99;
+    }
+
+    SamplingResult samplingResult = recordAndSampleWithSampleRateMap.get(sp);
     if (samplingResult == null) {
-      samplingResult = new RecordAndSampleWithItemCount(itemCount);
-      recordAndSampleWithItemCountMap.put(itemCount, samplingResult);
+      samplingResult = new RecordAndSampleWithItemCount(sp);
+      recordAndSampleWithSampleRateMap.put(sp, samplingResult);
     }
     return samplingResult;
   }
@@ -112,9 +125,10 @@ public class AiSampler implements Sampler {
       return SamplingResult.drop();
     }
     if (parentSpan instanceof ReadableSpan) {
-      Long itemCount = ((ReadableSpan) parentSpan).getAttribute(AiSemanticAttributes.ITEM_COUNT);
-      if (itemCount != null) {
-        return new RecordAndSampleWithItemCount(itemCount);
+      Double sampleRate =
+          ((ReadableSpan) parentSpan).getAttribute(AiSemanticAttributes.SAMPLE_RATE);
+      if (sampleRate != null) {
+        return new RecordAndSampleWithItemCount(sampleRate);
       }
     }
     return null;
@@ -141,8 +155,8 @@ public class AiSampler implements Sampler {
 
     private final Attributes attributes;
 
-    RecordAndSampleWithItemCount(long itemCount) {
-      attributes = Attributes.builder().put(AiSemanticAttributes.ITEM_COUNT, itemCount).build();
+    RecordAndSampleWithItemCount(double sampleRate) {
+      attributes = Attributes.builder().put(AiSemanticAttributes.SAMPLE_RATE, sampleRate).build();
     }
 
     @Override
